examples: reverted from AppiumOptions to well-known DesiredCapabilities.
AppiumOptions class was created to get aligned with future Appium 2.0 release.
However, as it's not exported with the SDK, being located under the test package.

Anyone who tries to use the example tests code, gets compilation errors as this class is missing.
To resolve this, examples were reverted back to the well-known DesiredCapabilities class.

Signed-off-by: Marat Strelets <marat@testproject.io>