fix: create generic driver reporter after creating the reporting command executor.
Create the reporting command executor before creating the reporter.
If running Cucumber, when creating the reporter, it will attempt to disable the auto test and command reports,
it will then cause a null pointer exception when the reporting command executor has not been created yet when
using a Generic driver.

Signed-off-by: david_goichman <david.goichman@testproject.io>